refactor: [FE] ResultSketchbook의 메인로직을 훅으로 분리
- Intent
변경될 기능을 추가하고 수정하기에는 따로 보는 것이 더 낫다고 생각해서 분리하였습니다.
